Fix Part of #277 : Added UrlImage Loading Tests (#5681)
**Explanation:**
Fix part of #277 by introducing a set of test cases to verify the
functionality of image loading via the `UrlImageParser` class using the
`GlideImageLoader`. The tests ensure that images (including PNG, SVG,
and SVGZ formats) are correctly loaded from URLs and that math drawables
(e.g., inline LaTeX) are rendered as expected. It covers scenarios for:
- Loading images from URLs of different formats (PNG, SVG, SVGZ).
- Verifying the behavior when loading multiple images at once.
- Ensuring that the image loading process finishes properly using Glide.

**Fixes:**
- This PR addresses the pending test cases for loading images from URLs
(as mentioned in issue #277) and resolves the related testing block.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [ ] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

**All Tests Passed:**

![Image](https://github.com/user-attachments/assets/df5c4015-1583-4faf-8f98-3d361c2b5bbc)