Fix #5749: Add support for checking lesson compatibility weekly (#5750)
## Explanation
Fixes #5749

This PR introduces support for, once per week (Mondays at 2:30am GMT),
attempting to download the list of lessons and failing if any conversion
errors occur (support for this was introduced in the asset download
script in 0252fc8fd5b246e56a3a34c88018eda2bf3a09f3 and
0bec9004cdb6faaa899d8b03162bce6b750d47de). There are currently over 6000
issues that need fixing, so it's expected that this is a cron job that
will be failing for some time as the content team works through the
failures, and the web team introduces checks to help prevent some of
these classes of failures from regressing.


https://github.com/BenHenning/oppia-android/actions/runs/13843084590/job/38735087365
demonstrates a failure when issues are detected:


![image](https://github.com/user-attachments/assets/ced33349-f83e-40bf-9613-368fdb8c1c1f)

Note that the actual exception thrown is caught as a GitHub actions
annotation, but fortunately it doesn't have any critically important
information as all the actual errors are printed to standard output.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- This is a CI infrastructure change.