Fix #5307: Remove all uses of ActivityScenarioRule. (#5755)
## Explanation
Fixes #5307

This migrates tests over to managing their `ActivityScenario` directly,
instead. This change does not affect existing uses of
`ActivityTestRule`.

The existing exemptions have been removed to ensure there are now zero
tests that are using `ActivityScenario`.

There were some additional clean-ups done, including:
- Removing extra lines for injections in a bunch of files.
- Adding a test target for `TextInputLayoutBindingAdaptersTest` since,
as a Java test, it isn't covered by the test file static check.
- The error for using `ActivityScenarioRule` was updated to drop the
platform parameters bit since, while true, it won't be as relevant after
#5725 is completed since it will become impossible to use pre-setup
orchestration of an activity.

Some other notes:
- This work is being done as a prerequisite to #5725 per
https://github.com/oppia/oppia-android/pull/5725#issuecomment-2727017259.
- `HtmlParserTest` required an extra bit of logic to reset a mock state
in
`testHtmlContent_withConceptCard_noLinkSupport_clickSpan_callsTagListener`
which is unexpected. Some other tests had slightly changed behaviors due
to the activity being initialized later, but this one actually had a
functional difference.
- Similarly, `testObserver_onFirstActivityResume_logsStartupLatency` in
`ApplicationLifecycleObserverTest` had an event order change, but this
seems less surprising since the change from rule to direct management is
definitely likely to impact some arrangement ordering as they are not
functionally equivalent.
- In a few cases, `ActivityScenarioRule` was being arranged but not used
(or ignored since a later explicit `launch` was used, anyway).
- I renamed `runWithActivity` to `onActivityWithResult` in a bunch of
places as a clearer, new pattern. The 'runWith' notion is really for the
`ActivityScenario`, now, and the actual operation being performed is an
`onActivity` with some sort of extraction (hence the 'withResult').
- Some extra work was needed in `MathExpressionAccessibilityUtilTest` to
avoid majorly changing a lot of tests (since the activity was needed
within a Truth assertion method). The change came out cleanly, though
something like Kotlin context receivers would be even cleaner (or their
new replacement).
- `ApplicationLifecycleObserverTest` had a bunch of module changes since
it was migrated over to using a more appropriate test activity fixture
(specifically `TestActivity`) that that required a lot more of the app
to be included as part of the build.
- `ApplicationLifecycleObserver` was exempted from code coverage since
it seems that the new modules being brought into the test are triggering
some of the duplicate instrumentation issues that we see with a lot of
app layer tests.
- Despite the length of the PR, much of these changes are space-only due
to new indentations being added in a few hundred tests.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- This only affects tests.