Fix #5308: Remove uses of ActivityTestRule. (#5759)
## Explanation
Fixes #5308

Similar to #5755, this removes all uses of `ActivityTestRule`.

Broadly, this was done by:
- Replacing `ActivityTestRule` uses with equivalent 'launch' calls with
`ActivityScenario`. This was doable in many places, and generally
required little to no functional differences in the test (just some
`onActivity` calls).
- Replacing all uses of `ActivityScenario` and `ActivityTestRule` with a
'runWith*' pattern similar to `ProfileAndDeviceIdFragmentTest`:
https://github.com/oppia/oppia-android/blob/b68f993733cc9a9c830dc034c3fc9a649bd32bd5/app/src/sharedTest/java/org/oppia/android/app/administratorcontrols/learneranalytics/ProfileAndDeviceIdFragmentTest.kt#L932-L947
(also similar to what was done in #5755). Note that this was done in a
few test suites where it had the most value (or where there was zero use
of `ActivityScenario` already).
- Continuing the pattern of relying on `TestActivity` when testing
fragments and views (where possible) rather than an actual activity
since this better validates the fragment's API rather than the
activity's behaviors. Callback listeners are verified with custom
extensions of `TestActivity` that delegate their calls to mocks that can
be verified using Mockito's utilities.

There were separate changes made or things to note:
- Some of the unnecessary casting to parent activities in fragments were
removed when finishing the activity. This is something we should
probably clean up broadly, but that wasn't done in this PR (generally,
it was only done were necessary or where noticed when investigating test
failures that might have been related).
- `TestActivity` was updated to include some helpers for registering
test activities without requiring an `AndroidManifest.xml` registration.
The issue is documented at
https://github.com/robolectric/robolectric/pull/4736 and this will help
us actually remove all of our test activities in the future (per #3830).
- Test injections and rule declarations were simplified and condensed
for better readability.
- `LessonThumbnailImageViewTest` had its tests largely redone to avoid
the need for a custom `LessonThumbnailImageViewTestActivity` (and test
fragment). These classes aren't being removed in this PR, but they
should be now obsolete.
- In some cases, the screen name tests were reorganized a bit to make
them easier to read and to inline, where possible, "create intent"
helpers since they were largely not needed and could be simplified
(especially when launching using `ActivityScenario`).
- In most cases the 'runWith*' functions have a
`testCoroutineDispatchers.runCurrent()` before running the test except
for `TopicFragmentTest` and `ExplorationActivityTest` since some tests
were highly sensitive to arranging specific state before running
coroutines. In these cases the previous test behaviors were attempted to
be retained with regards to background task and state synchronization.
- `setUpTestApplicationComponent` was moved below the tests where it was
noticed as not already there for consistency with the vast majority of
test suites.
- `RecentlyPlayedActivity` has a new test suite that has moved over
tests from `RecentlyPlayedFragmentTest` since it was verifying
activity-specific behavior. Test names have been renamed accordingly.
- `TopicFragmentTest` has some behavior differences in on the test
fragment is being set up (large differences in that it's no longer
hosted in `TopicActivity` for these tests). Some tests had to be set up
to not pass a story ID in order to be made to work, but most were able
to be moved over to that approach for a bit more setup consistency.
- Most of the changes in `ExplorationActivityTest` are spacing related,
but due to the specific diffing behavior many of the lines are showing
as re-written. This is likely due to `startPlayingNewExploration` being
folded into a new "launch activity & start exploration" helper to reduce
on the redundancy in a lot of the tests.

Finally, similar to #5755 this change is also a prerequisite to #5725.

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- This is only affected tests and should have no impact on
user-facing flows.