Fix #77, part of #59: Revise testability of Retrofit services (#5806)
## Explanation

Fixes #77
Fixes part of #59 (as part of an ongoing PR chain, see #5804)

This PR does two batches of work:
1. It splits `NetworkModule` into distinct modules: `RetrofitModule` and
`RetrofitServiceModule`.
2. It replaces `MockPlatformParameterService`,
`MockFeedbackReportingService`, and `MockRetrofit` with orchestrator
alternatives that leverage OkHttp's `MockWebServer`, instead.

(1) is needed for the ongoing PR chain to introduce bundle modules and
consolidate on common test modules and configurations (leading up to
#5804 and #5805). It was discovered that decoupling the different
dependencies for OkHttp and Retrofit was a bit too cumbersome to be
practical in those PRs, so this one was introduced, instead.

As part of (1), it became clear that managing what appeared to be two
distinct means of testing HTTP endpoints (a mock Retrofit service and
`MockWebServer`) was tricky and inconsistent, so this PR pushes
everything over to just using `MockWebServer` as that provides the best
integration pathway (it ensures that the generated Retrofit services and
corresponding Moshi data objects are actually set up correctly, and that
calling code into OkHttp is correct--which is especially important for
resolving #2951).

From a design perspective, the new orchestrators are meant to simulate
the Oppia web backend by fixing specific responses with an assumed
deterministic order to depending on them.
`PlatformParameterIntegrationTest` does a good job showing this working
all the way to a UI that depends on the parameters. There's some
repetition between the two orchestrators that might be worth solving
with a common base class in the future (or leveraging some sort of
composition or mixin pattern), but this seems like a reasonable start.
The orchestrators were designed to keep relative parity with the mock
services that they are essentially replacing.

Some things to note:
- `SplashTestActivity` was updated to use `TestActivity` and
self-register itself (which means that it could be removed from the main
app manifest & its accessibility exemption removed). I didn't actually
intend to use programmatic registration, but at first I didn't
understand where this failure was coming from. I eventually discovered
that the package manager setup in `NetworkConfigTestModule` was
resetting already registered activities and later fixed this (plus added
a test for it), but I decided to keep the self-registration for
`SplashTestActivity` since that's in-line with long-term test activity
goals.
- There are some `post()` exemptions added for the new tests--this is an
unfortunate side effect of using simple regex to detect bad patterns.
These aren't actually the same `post()` calls that we want to try and
prevent, but the exemptions are needed regardless.
- `NetworkApiKey` was moved to `network_config_annotations` since it's
an annotation (it probably should have been in that library originally).
- The Dagger `NetworkApiKey` provider is now part of
`NetworkConfigProdModule` so that its counterpart
`NetworkConfigTestModule` can provide a non-prod version of the key (and
ditto for the instrumentation app).
- `Moshi` is now being injected as well since it has additional
configurations (due to the introduction of `GaePlatformParameterValue`).
This is mainly for the benefit of tests, but it wouldn't be harmful if
other parts of the codebase depended on this object if they needed to.
We may want to qualify it to be GAE-specific in the future.
- `ApiMockLoader` was removed in favor of just loading mocks directly
(since Gradle is no longer supported). Similarly, due to the mock
services being removed everything under
`testing/src/main/java/org/oppia/android/testing/network` was able to be
removed (which is a nice win to move things out of the top-level testing
directory).
- Note that the change to load mocks directly required some small
differences in asset management, i.e. an introduction of a test-only
assets library for the data layer.
- Some field names in `GaeUserSuppliedFeedback` were updated to match
their mock JSON counterparts (these aren't actually used yet since
feedback reporting hasn't been completed--see #5333).
- `GaePlatformParameterValue` was introduced to simplify parameter
deserialization (since Moshi can handle this quite cleanly with a custom
`JsonAdapter` rather than needing to declare an `Any` type and risk bugs
or crashes at the time of using the data). A test was added for this
class to help establish the pattern of testing Moshi objects moving
forward.
- There's some extra Robolectric initialization happening in
`NetworkConfigTestModule` to simplify test arrangement (which can be
seen several times in the PR when similar-looking package manager setup
code has been removed). This may not end up being the permanent home for
this, but it's a very reasonable place for it to live for now.
- There's some risk in the time-based orchestrator tests. These didn't
have a better option for now (at least until #2951 is solved) since
testing for a disconnect basically requires having a short response
timeout, and the request is always blocking since OkHttp isn't yet
configured to be thread cooperative with the rest of the app.
- Some `MockRetrofit` comments were removed in tests since they weren't
actually using `MockRetrofit` (as intended, and as we want long-term).
Note that `ApplicationLifecycleObserverTest` _was_ updated to use the
new `NetworkConfigTestModule` (in-line with this idea of centralizing
test modules).

## Essential Checklist
- [x] The PR title and explanation each start with "Fix #bugnum: " (If
this PR fixes part of an issue, prefix the title with "Fix part of
#bugnum: ...".)
- [x] Any changes to
[scripts/assets](https://github.com/oppia/oppia-android/tree/develop/scripts/assets)
files have their rationale included in the PR explanation.
- [x] The PR follows the [style
guide](https://github.com/oppia/oppia-android/wiki/Coding-style-guide).
- [x] The PR does not contain any unnecessary code changes from Android
Studio
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#undo-unnecessary-changes)).
- [x] The PR is made from a branch that's **not** called "develop" and
is up-to-date with "develop".
- [x] The PR is **assigned** to the appropriate reviewers
([reference](https://github.com/oppia/oppia-android/wiki/Guidance-on-submitting-a-PR#clarification-regarding-assignees-and-reviewers-section)).

## For UI-specific PRs only
N/A -- This should only affect testing infrastructure and test
behaviors.